Client/client_requirements.txt
pillow
opencv-python
